Fix sync tests
